requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytest-django==4.5.2
pytest==7.4.3